Validate the form field
we cannot validate the form feild by clicking the add button as usual. because we add preventDefautlt on it so it doesnt work like a submit button.
so we added an if statement to check if the values of bookTitle and bookAuthor are entered. otherwise alert to say enter it.